PaxePacketTest fixes

+++ trex-paxe/src/main/java/com/github/trex_paxos/paxe/NodeId.java
@@ -1,13 +1,13 @@
 package com.github.trex_paxos.paxe;
 
-public record NodeId(byte value) implements Comparable<NodeId> {
+public record NodeId(short value) implements Comparable<NodeId> {
   public NodeId {
       if (value < 0) throw new IllegalArgumentException("Node ID must be non-negative");
   }
   
   @Override
   public int compareTo(NodeId other) {
-      return Byte.compare(value, other.value);
+      return Short.compare(value, other.value);
   }
 
   @Override
